+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 * Added 1.19.0 support
 * Fixed a duplicate redirect warning
+* Fixed various bugs with the Project Table and Rafter